fix dependencies
