+++ packages/core-ethereum/src/ethereum.ts
@@ -2,9 +2,9 @@ import type { ContractTransaction } from 'ethers'
 import type Multiaddr from 'multiaddr'
 import type { HoprToken, HoprChannels } from './contracts'
 import { providers, utils, errors, Wallet, BigNumber } from 'ethers'
-import { Address, Ticket, Acknowledgement } from './types'
+import { Address, Ticket, Acknowledgement } from '@hoprnet/hopr-utils'
 import BN from 'bn.js'
-import { Balance, NativeBalance, Hash, PublicKey } from './types'
+import { Balance, NativeBalance, Hash, PublicKey } from '@hoprnet/hopr-utils'
 import { durations, PromiseValue } from '@hoprnet/hopr-utils'
 import NonceTracker from './nonce-tracker'
 import TransactionManager from './transaction-manager'
