+++ packages/core-ethereum/src/indexer/fixtures.ts
@@ -3,7 +3,7 @@ import assert from 'assert'
 import Multiaddr from 'multiaddr'
 import { BigNumber } from 'ethers'
 import { stringToU8a } from '@hoprnet/hopr-utils'
-import { PublicKey, Hash, AccountEntry, ChannelEntry } from '../types'
+import { PublicKey, Hash, AccountEntry, ChannelEntry } from '@hoprnet/hopr-utils'
 
 export const partyA = PublicKey.fromString('0x03362b7b26bddb151a03056422d37119eab3a716562b6c3efdc62dec1540c9b091')
 export const partyB = PublicKey.fromString('0x03217f3cd4d0b4b82997b25d1b6b68a933929fed724531cb30bbfd4729dc6b44e0')
