+++ packages/core-ethereum/src/indexer/index.ts
@@ -5,7 +5,7 @@ import chalk from 'chalk'
 import BN from 'bn.js'
 import Heap from 'heap-js'
 import { pubKeyToPeerId, randomChoice, HoprDB } from '@hoprnet/hopr-utils'
-import { Address, ChannelEntry, AccountEntry, Hash, PublicKey, Balance, Snapshot } from '../types'
+import { Address, ChannelEntry, AccountEntry, Hash, PublicKey, Balance, Snapshot } from '@hoprnet/hopr-utils'
 import { isConfirmedBlock, snapshotComparator } from './utils'
 import Debug from 'debug'
 import Multiaddr from 'multiaddr'
