+++ packages/core/src/messages/acknowledgement/index.ts
@@ -2,7 +2,7 @@ import secp256k1 from 'secp256k1'
 import { deriveTicketKeyBlinding } from '../packet/header'
 import { KEY_LENGTH } from '../packet/header/parameters'
 import { Challenge } from '../packet/challenge'
-import { Hash, Signature, PublicKey } from '@hoprnet/hopr-core-ethereum'
+import { Hash, Signature, PublicKey } from '@hoprnet/hopr-utils'
 import PeerId from 'peer-id'
 import { serializeToU8a, u8aSplit, u8aToHex, UnAcknowledgedTickets } from '@hoprnet/hopr-utils'
