+++ packages/core/src/network/heartbeat.ts
@@ -1,7 +1,7 @@
 import type NetworkPeerStore from './network-peers'
 import type PeerId from 'peer-id'
 import debug from 'debug'
-import { Hash } from '@hoprnet/hopr-core-ethereum'
+import { Hash } from '@hoprnet/hopr-utils'
 import { randomInteger, limitConcurrency, LibP2PHandlerFunction, u8aEquals, DialOpts } from '@hoprnet/hopr-utils'
 import { HEARTBEAT_INTERVAL, HEARTBEAT_INTERVAL_VARIANCE, MAX_PARALLEL_CONNECTIONS } from '../constants'
 import { PROTOCOL_HEARTBEAT, HEARTBEAT_TIMEOUT } from '../constants'
