+++ packages/hoprd/src/commands/fundChannel.ts
@@ -2,8 +2,7 @@ import type Hopr from '@hoprnet/hopr-core'
 import type PeerId from 'peer-id'
 import chalk from 'chalk'
 import BN from 'bn.js'
-import { moveDecimalPoint } from '@hoprnet/hopr-utils'
-import { Balance } from '@hoprnet/hopr-core-ethereum'
+import { moveDecimalPoint, Balance } from '@hoprnet/hopr-utils'
 import { AbstractCommand, GlobalState } from './abstractCommand'
 import { checkPeerIdInput, styleValue } from './utils'
