+++ packages/hoprd/src/commands/withdraw.ts
@@ -2,7 +2,7 @@ import type Hopr from '@hoprnet/hopr-core'
 import { moveDecimalPoint } from '@hoprnet/hopr-utils'
 import { AbstractCommand, AutoCompleteResult } from './abstractCommand'
 import { styleValue } from './utils'
-import { Balance, NativeBalance } from '@hoprnet/hopr-core-ethereum'
+import { Balance, NativeBalance } from '@hoprnet/hopr-utils'
 
 export default class Withdraw extends AbstractCommand {
   private arguments = ['amount (ETH, HOPR)', 'currency (native, hopr)', 'recipient (blockchain address)']
