+++ packages/utils/src/types/accountEntry.ts
@@ -1,4 +1,3 @@
-import type { Event } from './indexer'
 import Multiaddr from 'multiaddr'
 import PeerId from 'peer-id'
 import { ethers } from 'ethers'
@@ -23,7 +22,7 @@ class AccountEntry {
     return new AccountEntry(address, isBEmpty ? undefined : Multiaddr(strippedB))
   }
 
-  static fromSCEvent(event: Event<'Announcement'>): AccountEntry {
+  static fromSCEvent(event: any): AccountEntry { //TODO types
     const { account, multiaddr } = event.args
     const address = Address.fromString(account)
     const accountEntry = new AccountEntry(address, Multiaddr(multiaddr))
